# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenLedger API
  version: 0.1.0
  description: >-
    OIDC-authenticated account and internal transfer API. Customer funds move only
    through balanced transactions committed by the ledger database.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness and Postgres connectivity
      responses:
        '200':
          description: Process is up and can query Postgres
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Process is up but Postgres is unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/accounts:
    post:
      operationId: createAccount
      summary: Create a customer account for the authenticated principal
      security:
        - OIDCBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/accounts/{id}:
    get:
      operationId: getAccount
      summary: Get an account owned by the authenticated principal
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/accounts/{id}/balance:
    get:
      operationId: getAccountBalance
      summary: Get an account balance owned by the authenticated principal
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Balance in exact integer minor units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/transfers:
    post:
      operationId: createTransfer
      summary: Transfer funds between owned customer accounts
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '200':
          description: Previously committed result returned for an idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '201':
          description: Transfer posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Insufficient funds, currency mismatch, or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/transfers/{id}:
    get:
      operationId: getTransfer
      summary: Get a transfer involving accounts owned by the authenticated principal
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/deposits:
    post:
      operationId: createDepositIntent
      summary: Create an authenticated Paystack deposit intent
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepositRequest'
      responses:
        '202':
          description: Payment intent queued for Paystack initialization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositAccepted'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Provider integration is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/deposits/{reference}:
    get:
      operationId: getDepositIntent
      summary: Get a deposit intent owned by the authenticated principal
      security:
        - OIDCBearer: []
      parameters:
        - $ref: '#/components/parameters/PaymentReference'
      responses:
        '200':
          description: Deposit payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositPayment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /webhooks/paystack:
    post:
      operationId: receivePaystackWebhook
      summary: Receive and deduplicate signed Paystack webhook events
      parameters:
        - $ref: '#/components/parameters/PaystackSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaystackWebhook'
      responses:
        '202':
          description: Signed event durably queued for reconciliation
        '200':
          description: Duplicate or unsupported event acknowledged
        '400':
          description: Invalid event body
        '401':
          description: Invalid signature
        '409':
          description: Reused provider event key has a different payload
components:
  securitySchemes:
    OIDCBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    PaymentReference:
      name: reference
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    PaystackSignature:
      name: x-paystack-signature
      in: header
      required: true
      description: HMAC-SHA512 of the exact raw request body using the Paystack secret key.
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{128}$'
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Required ledger scope is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or not owned by the authenticated principal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Currency:
      type: string
      enum: [NGN, USD]
    MinorUnitsValue:
      oneOf:
        - type: integer
          minimum: 0
          maximum: 9007199254740991
        - type: string
          pattern: '^(0|[1-9][0-9]*)$'
      description: Exact integer minor units; values beyond JavaScript safe integer range are decimal strings.
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, postgres]
      properties:
        status:
          type: string
          enum: [ok]
        postgres:
          type: string
          enum: [up]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [currency]
      properties:
        currency:
          $ref: '#/components/schemas/Currency'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
    Account:
      type: object
      additionalProperties: false
      required: [id, currency, displayName, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        currency:
          $ref: '#/components/schemas/Currency'
        displayName:
          type: string
        status:
          type: string
          enum: [active, frozen, closed]
        createdAt:
          type: string
          format: date-time
    Balance:
      type: object
      additionalProperties: false
      required: [accountId, amountMinor, currency]
      properties:
        accountId:
          type: string
          format: uuid
        amountMinor:
          $ref: '#/components/schemas/MinorUnitsValue'
        currency:
          $ref: '#/components/schemas/Currency'
    CreateTransferRequest:
      type: object
      additionalProperties: false
      required: [sourceAccountId, destinationAccountId, amountMinor, currency]
      properties:
        sourceAccountId:
          type: string
          format: uuid
        destinationAccountId:
          type: string
          format: uuid
        amountMinor:
          type: integer
          minimum: 1
          maximum: 9007199254740991
          description: Positive integer minor units, within exact JSON integer range.
        currency:
          $ref: '#/components/schemas/Currency'
    CreateDepositRequest:
      type: object
      additionalProperties: false
      required: [accountId, amountMinor, currency, email]
      properties:
        accountId:
          type: string
          format: uuid
        amountMinor:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        currency:
          $ref: '#/components/schemas/Currency'
        email:
          type: string
          format: email
          maxLength: 254
    DepositAccepted:
      type: object
      additionalProperties: false
      required: [reference, status]
      properties:
        reference:
          type: string
        status:
          type: string
          enum: [pending]
    DepositPayment:
      type: object
      additionalProperties: false
      required: [reference, amountMinor, currency, status, authorizationUrl, createdAt]
      properties:
        reference:
          type: string
        amountMinor:
          $ref: '#/components/schemas/MinorUnitsValue'
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          type: string
          enum: [pending, initialized, succeeded, failed, reconciliation_required]
        authorizationUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaystackWebhook:
      type: object
      required: [event, data]
      properties:
        event:
          type: string
        data:
          type: object
          required: [reference, amount, currency, status]
          properties:
            reference:
              type: string
            amount:
              type: integer
              minimum: 1
            currency:
              $ref: '#/components/schemas/Currency'
            status:
              type: string
    Transfer:
      type: object
      additionalProperties: false
      required: [id, sourceAccountId, destinationAccountId, amountMinor, currency, createdAt]
      properties:
        id:
          type: string
          format: uuid
        sourceAccountId:
          type: string
          format: uuid
        destinationAccountId:
          type: string
          format: uuid
        amountMinor:
          $ref: '#/components/schemas/MinorUnitsValue'
        currency:
          $ref: '#/components/schemas/Currency'
        createdAt:
          type: string
          format: date-time